Move UploadedFile and ImportRecipesResult to domain layer
These dataclasses are shared across multiple usecases (the existing
EXIF-import and a forthcoming QR-card-import), so they belong in the
domain layer rather than inside one of the usecase modules.

No behaviour change.
